Moved ChildHealthMonthlyAggregationHelper to own file
